feat: add designer controller

+++ src/main/java/robert/swag/pwebbe/controllers/MaterialController.java
@@ -23,21 +23,43 @@ public class MaterialController {
         return ResponseEntity.ok().body(materialService.get());
     }
 
+    @GetMapping("/{id}")
+    @PreAuthorize("hasAuthority('USER')")
+    public ResponseEntity<Material> getById(@PathVariable Long id) {
+        if (!materialService.verifyMaterialExists(id)) {
+            return ResponseEntity.status(404).body(null);
+        }
+
+        return ResponseEntity.ok().body(materialService.getById(id));
+    }
+
     @PostMapping
     @PreAuthorize("hasAuthority('USER')")
     public ResponseEntity<Material> add(@RequestBody MaterialDto material) {
+        if (materialService.verifyMaterialExists(material.getName())) {
+            return ResponseEntity.status(400).body(null);
+        }
+
         return ResponseEntity.ok().body(materialService.add(material));
     }
 
     @PostMapping("/update")
     @PreAuthorize("hasAuthority('USER')")
     public ResponseEntity<Material> update(@RequestBody MaterialDto materialDto) {
+        if (materialDto.getId() == null || !materialService.verifyMaterialExists(materialDto.getId())) {
+            return ResponseEntity.status(404).body(null);
+        }
+
         return ResponseEntity.ok().body(materialService.update(materialDto));
     }
 
     @DeleteMapping("/{id}")
-    @PreAuthorize("hasAuthority('USER')")
+    @PreAuthorize("hasAuthority('ADMIN')")
     public ResponseEntity<Void> delete(@PathVariable Long id) {
+        if (!materialService.verifyMaterialExists(id)) {
+            return ResponseEntity.status(404).body(null);
+        }
+
         materialService.delete(id);
         return ResponseEntity.ok().body(null);
     }
